oceanmgr/requirements.txt
python-dotenv
requests
#sqlite3
